separated out ClientTester
